+++ hata/discord/activity/metadata/tests/test__CustomActivityMetadata__data.py
@@ -5,26 +5,26 @@ import vampytest
 from ....emoji import Emoji, create_partial_emoji_data
 from ....utils import datetime_to_millisecond_unix_time
 
-from .. import CustomActivityMetadata
+from .. import ActivityMetadataCustom
 
 
-def test__CustomActivityMetadata__from_data__0():
+def test__ActivityMetadataCustom__from_data__0():
     """
-    Tests whether ``CustomActivityMetadata.from_data`` works as intended.
+    Tests whether ``ActivityMetadataCustom.from_data`` works as intended.
     
     Case: No fields given.
     """
-    activity_metadata = CustomActivityMetadata.from_data({})
+    activity_metadata = ActivityMetadataCustom.from_data({})
     
-    vampytest.assert_instance(activity_metadata, CustomActivityMetadata)
+    vampytest.assert_instance(activity_metadata, ActivityMetadataCustom)
     vampytest.assert_is(activity_metadata.state, None)
     vampytest.assert_is(activity_metadata.emoji, None)
     vampytest.assert_is(activity_metadata.created_at, None)
 
 
-def test__CustomActivityMetadata__from_data__1():
+def test__ActivityMetadataCustom__from_data__1():
     """
-    Tests whether ``CustomActivityMetadata.from_data`` works as intended.
+    Tests whether ``ActivityMetadataCustom.from_data`` works as intended.
     
     Case: All fields given.
     """
@@ -32,27 +32,27 @@ def test__CustomActivityMetadata__from_data__1():
     emoji = Emoji.precreate(202209060003, name='Code49')
     created_at = DateTime(2014, 9, 16)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': state,
         'emoji': create_partial_emoji_data(emoji),
         'created_at': datetime_to_millisecond_unix_time(created_at),
     })
     
-    vampytest.assert_instance(activity_metadata, CustomActivityMetadata)
+    vampytest.assert_instance(activity_metadata, ActivityMetadataCustom)
     vampytest.assert_eq(activity_metadata.state, state)
     vampytest.assert_is(activity_metadata.emoji, emoji)
     vampytest.assert_eq(activity_metadata.created_at, created_at)
 
 
-def test__CustomActivityMetadata__to_data():
+def test__ActivityMetadataCustom__to_data():
     """
-    Tests whether ``CustomActivityMetadata.to_data`` works as intended.
+    Tests whether ``ActivityMetadataCustom.to_data`` works as intended.
     """
     state = 'state'
     emoji = Emoji.precreate(202209060004, name='Code49')
     created_at = DateTime(2014, 9, 16)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': state,
         'emoji': create_partial_emoji_data(emoji),
         'created_at': datetime_to_millisecond_unix_time(created_at),
@@ -61,15 +61,15 @@ def test__CustomActivityMetadata__to_data():
     vampytest.assert_eq(activity_metadata.to_data(), {})
 
 
-def test__CustomActivityMetadata__to_data_user():
+def test__ActivityMetadataCustom__to_data_user():
     """
-    Tests whether ``CustomActivityMetadata.to_data_user`` works as intended.
+    Tests whether ``ActivityMetadataCustom.to_data_user`` works as intended.
     """
     state = 'state'
     emoji = Emoji.precreate(202209060005, name='Code49')
     created_at = DateTime(2014, 9, 16)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': state,
         'emoji': create_partial_emoji_data(emoji),
         'created_at': datetime_to_millisecond_unix_time(created_at),
@@ -78,15 +78,15 @@ def test__CustomActivityMetadata__to_data_user():
     vampytest.assert_eq(activity_metadata.to_data(), {})
 
 
-def test__CustomActivityMetadata__to_data_full():
+def test__ActivityMetadataCustom__to_data_full():
     """
-    Tests whether ``CustomActivityMetadata.to_data_full`` works as intended.
+    Tests whether ``ActivityMetadataCustom.to_data_full`` works as intended.
     """
     state = 'state'
     emoji = Emoji.precreate(202209060006, name='Code49')
     created_at = DateTime(2014, 9, 16)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': state,
         'emoji': create_partial_emoji_data(emoji),
         'created_at': datetime_to_millisecond_unix_time(created_at),
@@ -103,9 +103,9 @@ def test__CustomActivityMetadata__to_data_full():
     vampytest.assert_eq(data['created_at'], datetime_to_millisecond_unix_time(created_at))
 
 
-def test__CustomActivityMetadata__update_attributes():
+def test__ActivityMetadataCustom__update_attributes():
     """
-    Tests whether ``CustomActivityMetadata._update_attributes`` works as intended.
+    Tests whether ``ActivityMetadataCustom._update_attributes`` works as intended.
     """
     old_state = 'state'
     new_state = 'remilia'
@@ -114,7 +114,7 @@ def test__CustomActivityMetadata__update_attributes():
     old_created_at = DateTime(2014, 9, 16)
     new_created_at = DateTime(2014, 9, 17)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': old_state,
         'emoji': create_partial_emoji_data(old_emoji),
         'created_at': datetime_to_millisecond_unix_time(old_created_at),
@@ -131,9 +131,9 @@ def test__CustomActivityMetadata__update_attributes():
     vampytest.assert_eq(activity_metadata.created_at, new_created_at)
 
 
-def test__CustomActivityMetadata__difference_update_attributes():
+def test__ActivityMetadataCustom__difference_update_attributes():
     """
-    Tests whether ``CustomActivityMetadata._difference_update_attributes`` works as intended.
+    Tests whether ``ActivityMetadataCustom._difference_update_attributes`` works as intended.
     """
     old_state = 'state'
     new_state = 'remilia'
@@ -142,7 +142,7 @@ def test__CustomActivityMetadata__difference_update_attributes():
     old_created_at = DateTime(2014, 9, 16)
     new_created_at = DateTime(2014, 9, 17)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': old_state,
         'emoji': create_partial_emoji_data(old_emoji),
         'created_at': datetime_to_millisecond_unix_time(old_created_at),