+++ hata/discord/activity/metadata/tests/test__CustomActivityMetadata__magic.py
@@ -5,18 +5,18 @@ import vampytest
 from ....emoji import Emoji, create_partial_emoji_data
 from ....utils import datetime_to_millisecond_unix_time
 
-from .. import CustomActivityMetadata
+from .. import ActivityMetadataCustom
 
 
-def test__CustomActivityMetadata__repr():
+def test__ActivityMetadataCustom__repr():
     """
-    Tests whether ``CustomActivityMetadata.__repr__`` works as intended.
+    Tests whether ``ActivityMetadataCustom.__repr__`` works as intended.
     """
     state = 'state'
     emoji = Emoji.precreate(202209060000, name='Code49')
     created_at = DateTime(2014, 9, 16)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': state,
         'emoji': create_partial_emoji_data(emoji),
         'created_at': datetime_to_millisecond_unix_time(created_at),
@@ -25,15 +25,15 @@ def test__CustomActivityMetadata__repr():
     vampytest.assert_instance(repr(activity_metadata), str)
 
 
-def test__CustomActivityMetadata__hash():
+def test__ActivityMetadataCustom__hash():
     """
-    Tests whether ``CustomActivityMetadata.__hash__`` works as intended.
+    Tests whether ``ActivityMetadataCustom.__hash__`` works as intended.
     """
     state = 'state'
     emoji = Emoji.precreate(202209060001, name='Code49')
     created_at = DateTime(2014, 9, 16)
     
-    activity_metadata = CustomActivityMetadata.from_data({
+    activity_metadata = ActivityMetadataCustom.from_data({
         'state': state,
         'emoji': create_partial_emoji_data(emoji),
         'created_at': datetime_to_millisecond_unix_time(created_at),
@@ -42,9 +42,9 @@ def test__CustomActivityMetadata__hash():
     vampytest.assert_instance(hash(activity_metadata), int)
 
 
-def test__CustomActivityMetadata__eq():
+def test__ActivityMetadataCustom__eq():
     """
-    Tests whether ``CustomActivityMetadata.__eq__`` works as intended.
+    Tests whether ``ActivityMetadataCustom.__eq__`` works as intended.
     """
     state = 'state'
     emoji = Emoji.precreate(202209060002, name='Code49')
@@ -56,7 +56,7 @@ def test__CustomActivityMetadata__eq():
         'created_at': datetime_to_millisecond_unix_time(created_at),
     }
     
-    activity_metadata = CustomActivityMetadata.from_data(data)
+    activity_metadata = ActivityMetadataCustom.from_data(data)
     
     vampytest.assert_eq(activity_metadata, activity_metadata)
     vampytest.assert_ne(activity_metadata, object())
@@ -66,5 +66,5 @@ def test__CustomActivityMetadata__eq():
         'emoji',
         'created_at',
     ):
-        temporary_activity_metadata = CustomActivityMetadata.from_data({**data, field_name: None})
+        temporary_activity_metadata = ActivityMetadataCustom.from_data({**data, field_name: None})
         vampytest.assert_ne(temporary_activity_metadata, activity_metadata)